requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
